Add icons to git
There's a bug with the Tauri bundler where it freezes up if icon files
are missing. I cleaned the directory and forgot to regenerate, so this
happened to me. It's a bit cleaner to leave them out since the SVG and
`generate` script can rebuild them, but it's not worth having to
remember that quirk. All the Tauri examples keep the icons in the repo,
so perhaps they just don't support what I tried to do.